refactor(nifs): mv `Relaxed*` into 'nifs::vanilla'
**Motivation**
After the introduction of new types `FoldingScheme` `Relaxed*`
structures turned out to be just one of the subtypes of accumulators and
to keep their fields out of the way, they were moved to a separate
structure

**Overview**
No new code was introduced as part of this PR, just moved the strings
and renamed the `to_relax` -> `from_regular` methods
