+++ src/nifs/protogalaxy/accumulator.rs
@@ -3,7 +3,7 @@ use std::iter;
 use crate::{
     ff::Field,
     halo2curves::CurveAffine,
-    plonk::{RelaxedPlonkInstance, RelaxedPlonkTrace, RelaxedPlonkTraceArgs},
+    plonk::{self, PlonkInstance, PlonkTrace},
     poseidon::{AbsorbInRO, ROTrait},
     util,
 };
@@ -13,7 +13,7 @@ use crate::{
 pub struct Accumulator<C: CurveAffine> {
     /// `φ`: Represents the combined state of all instances & witnesses. It is a summary that
     /// captures the essential data and relationships from the instances being merged.
-    pub(super) trace: RelaxedPlonkTrace<C>,
+    pub(super) trace: PlonkTrace<C>,
 
     /// `β`: A random value used in the folding process. It helps ensure the unique
     /// and secure combination of instances, preventing manipulation.
@@ -26,7 +26,7 @@ pub struct Accumulator<C: CurveAffine> {
 
 impl<C: CurveAffine, RO: ROTrait<C::Base>> AbsorbInRO<C::Base, RO> for Accumulator<C> {
     fn absorb_into(&self, ro: &mut RO) {
-        ro.absorb(&self.trace.U).absorb_field_iter(
+        ro.absorb(&self.trace.u).absorb_field_iter(
             self.betas
                 .iter()
                 .chain(iter::once(&self.e))
@@ -35,14 +35,14 @@ impl<C: CurveAffine, RO: ROTrait<C::Base>> AbsorbInRO<C::Base, RO> for Accumulat
     }
 }
 
-pub type AccumulatorArgs = RelaxedPlonkTraceArgs;
+pub type AccumulatorArgs = plonk::PlonkTraceArgs;
 
 impl<C: CurveAffine> Accumulator<C> {
     pub fn new(args: AccumulatorArgs, count_of_evaluation: usize) -> Self {
         Self {
             betas: vec![C::ScalarExt::ZERO; count_of_evaluation].into_boxed_slice(),
             e: C::ScalarExt::ZERO,
-            trace: RelaxedPlonkTrace::new(args),
+            trace: PlonkTrace::new(args),
         }
     }
 }
@@ -52,7 +52,7 @@ impl<C: CurveAffine> Accumulator<C> {
 pub struct AccumulatorInstance<C: CurveAffine> {
     /// `φ`: Represents the combined state of all instances. It is a summary that captures the
     /// essential data and relationships from the instances being merged.
-    pub(super) U: RelaxedPlonkInstance<C>,
+    pub(super) ins: PlonkInstance<C>,
 
     /// `β`: A random value used in the folding process. It helps ensure the unique
     /// and secure combination of instances, preventing manipulation.
@@ -67,17 +67,17 @@ impl<C: CurveAffine> From<Accumulator<C>> for AccumulatorInstance<C> {
     fn from(value: Accumulator<C>) -> Self {
         let Accumulator {
             betas,
-            trace: RelaxedPlonkTrace { U, W: _ },
+            trace: PlonkTrace { u: ins, w: _ },
             e,
         } = value;
 
-        AccumulatorInstance { betas, U, e }
+        AccumulatorInstance { betas, ins, e }
     }
 }
 
 impl<C: CurveAffine, RO: ROTrait<C::Base>> AbsorbInRO<C::Base, RO> for AccumulatorInstance<C> {
     fn absorb_into(&self, ro: &mut RO) {
-        ro.absorb(&self.U).absorb_field_iter(
+        ro.absorb(&self.ins).absorb_field_iter(
             self.betas
                 .iter()
                 .chain(iter::once(&self.e))
